Bridge between population and Simulation
Within the simulation class, it simulates the spread of COVID-19 by manipulating the raw data of the population directly. Moreover, it also conditions the destination and many other configurations, and interacts with many functions outer the file,  even making the main body of the simulation harder to read and edit. 
To cope with this problem, or more specifically, to deal with the complex manipulation on population and the messed calling of outer functions, the Bridge pattern is implemented, and the raw data of population, as well as the corresponding manipulation process, are encapsulated in class SimPopulation. Then, with SimPopulation, Simulation no longer needs to access the raw data directly. Instead, it can just call the public function in SimPopulation, letting us more convenient to manage the actions and data during each iteration.